Change the data type of attribute value in structure trap_point from uint8_t to uint64_t

+++ common/include/defs.h
@@ -82,7 +82,7 @@ struct trap_point {
 
   /* Byte that was overwritten by the int3, needed so we can overwrite and
    * execute the original instruction */
-  uint8_t value;
+  uint64_t value;
 
   /* Index into the function array for the containing function */
   int fcn_i;
